backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydantic-settings>=2.2.0
pandas>=2.0.0
openpyxl>=3.1.0
gunicorn>=22.0.0
